- Split test_models.py into test_menu_classes.py and test_page_models.py - Move the 'TestNoAbsoluteUrlsPage' test case into test_custom_models.py, rather than having it in 'test_page_models.py', which is intended for testing the model functionality from wagtailmenus/models/pages.py than from custom models
